+++ src/microstructpy/examples/foam.py
@@ -1,8 +1,8 @@
 import os
 
-import matplotlib.pyplot as plt
 import numpy as np
 import scipy.stats
+from matplotlib import pyplot as plt
 
 import microstructpy as msp